Extract endpoints to file
